Полный и тщательный рефакторинг и реформат всего проекта

+++ src/PhotoCatalog.Application/UseCases/DeletePhotoUseCase.cs
@@ -1,5 +1,3 @@
-using Microsoft.Extensions.Logging;
-
 using PhotoCatalog.Application.Errors;
 using PhotoCatalog.Domain.Entities;
 using PhotoCatalog.Domain.Extensions;
@@ -7,21 +5,24 @@ using PhotoCatalog.Domain.Interfaces.Repositories;
 using PhotoCatalog.Domain.Interfaces.Services;
 using PhotoCatalog.Domain.Primitives;
 
+using Serilog;
+
 namespace PhotoCatalog.Application.UseCases;
 
 /// <summary>
 ///     Представляет прикладную сущность для удаления файла из репозитория и диска.
 /// </summary>
-/// <param name="photoRepository">репозиторий фотографий.</param>
-/// <param name="fileStorage">хранение файлов.</param>
-/// <param name="unitOfWork">единица работы.</param>
-/// <param name="logger">логгер.</param>
+/// <param name="photoQueryRepository">Репозиторий фотографий для получения данных.</param>
+/// <param name="photoCommandRepository">Репозиторий фотографий для добавления данных.</param>
+/// <param name="fileStorage">Хранилище файлов.</param>
+/// <param name="unitOfWork">Единица работы.</param>
+/// <param name="logger">Логгер.</param>
 public class DeletePhotoUseCase(
     IPhotoQueryRepository photoQueryRepository,
     IPhotoCommandRepository photoCommandRepository,
     IFileStorage fileStorage,
     IUnitOfWork unitOfWork,
-    ILogger<DeletePhotoUseCase> logger)
+    ILogger logger)
 {
     /// <summary>
     ///     Метод для удаления файла в репозитории и на диске по идентификатору.
@@ -37,18 +38,23 @@ public class DeletePhotoUseCase(
             .ToResult(ApplicationErrors.General.NotFound)
             .Value;
 
+        if (photo is { Value: null } or null)
+        {
+            return ResultVoid.Failure(ApplicationErrors.General.NotFound);
+        }
+
         unitOfWork.BeginTransaction();
 
         photoCommandRepository.Delete(photoId);
 
+
         if (unitOfWork.Commit().IsSuccess)
         {
-            fileStorage.DeleteFile(photo!.Value!.RealPath);
-
+            fileStorage.DeleteFile(photo.Value.RealPath);
             return ResultVoid.Success();
         }
 
-        logger.LogError("Orphaned file left on disk: {Path}", photo!.Value!.RealPath);
+        logger.Error("Orphaned file left on disk: {Path}", photo.Value.RealPath);
 
         return ResultVoid.Failure(ApplicationErrors.Files.OrphanedFile);
     }
